requirements.txt
Flask
prometheus-client
pyserial
gpiozero
RPi.GPIO